requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
onnx==1.20.1
onnxruntime-gpu==1.24.4
onnxslim==0.1.89
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
polars==1.39.3
protobuf==7.34.0
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
six==1.17.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
ultralytics==8.4.24
ultralytics-thop==2.0.18
urllib3==2.6.3
tensorrt-cu12>=7.0.0,!=10.1.0
